+++ controller/src/main/java/ru/hikkiyomi/dtos/SimpleKitten.java
@@ -1,3 +1,3 @@
 package ru.hikkiyomi.dtos;
 
-public record SimpleKitten(int id, String name) { }
+public record SimpleKitten(Long id, String name) { }
