+++ controller/src/main/java/ru/hikkiyomi/dtos/SimpleOwner.java
@@ -1,3 +1,3 @@
 package ru.hikkiyomi.dtos;
 
-public record SimpleOwner(int id, String name) { }
+public record SimpleOwner(Long id, String name) { }
